Extract email verification error handling to hook
